+++ pandas/tests/frame/indexing/test_setitem.py
@@ -10,10 +10,12 @@ from pandas import (
     Interval,
     NaT,
     Period,
+    PeriodIndex,
     Series,
     Timestamp,
     date_range,
     notna,
+    period_range,
 )
 import pandas._testing as tm
 from pandas.core.arrays import SparseArray
@@ -213,3 +215,17 @@ class TestDataFrameSetItem:
         result = df2["B"]
         tm.assert_series_equal(notna(result), Series([True, False, True], name="B"))
         tm.assert_series_equal(df2.dtypes, df.dtypes)
+
+    def test_setitem_periodindex(self):
+        rng = period_range("1/1/2000", periods=5, name="index")
+        df = DataFrame(np.random.randn(5, 3), index=rng)
+
+        df["Index"] = rng
+        rs = Index(df["Index"])
+        tm.assert_index_equal(rs, rng, check_names=False)
+        assert rs.name == "Index"
+        assert rng.name == "index"
+
+        rs = df.reset_index().set_index("index")
+        assert isinstance(rs.index, PeriodIndex)
+        tm.assert_index_equal(rs.index, rng)
