+++ pandas/tests/frame/test_period.py
@@ -1,6 +1,6 @@
 import numpy as np
 
-from pandas import DataFrame, Index, PeriodIndex, period_range
+from pandas import DataFrame, period_range
 import pandas._testing as tm
 
 
@@ -17,17 +17,3 @@ class TestPeriodIndex:
 
         ts = df["1/1/2000"]
         tm.assert_series_equal(ts, df.iloc[:, 0])
-
-    def test_frame_setitem(self):
-        rng = period_range("1/1/2000", periods=5, name="index")
-        df = DataFrame(np.random.randn(5, 3), index=rng)
-
-        df["Index"] = rng
-        rs = Index(df["Index"])
-        tm.assert_index_equal(rs, rng, check_names=False)
-        assert rs.name == "Index"
-        assert rng.name == "index"
-
-        rs = df.reset_index().set_index("index")
-        assert isinstance(rs.index, PeriodIndex)
-        tm.assert_index_equal(rs.index, rng)
